+++ tests/testthat/test-make.metric.R
@@ -81,6 +81,17 @@ test_that("Output is correct", {
     	make.metric(function(x)sd(variances(var(x))), silent=TRUE), "level1"
     	)
 
+    ## Same with data.dim
+    expect_equal(
+        make.metric(function(x)variances(var(x)), silent=TRUE, data.dim = c(3, 2)), "level2"
+        )
+    expect_equal(
+        make.metric(function(x)var(var(x)), silent=TRUE, data.dim = c(33, 5)), "level3"
+        )
+    expect_equal(
+        make.metric(function(x)sd(variances(var(x))), silent=TRUE, data.dim = c(3, 10)), "level1"
+        )
+
     expect_equal(capture.output(make.metric(var)),
         c("var outputs a matrix object.", "var is detected as being a dimension-level 3 function.", "Additional dimension-level 2 and/or 1 function(s) will be needed."))